hackadoll1/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
-e git+https://github.com/tsukumijima/youtube-dl.git@fork#egg=youtube-dl  # via -r requirements.in
aiohttp==3.7.4            # via discord-py
async-timeout==3.0.1      # via aiohttp
attrs==20.3.0             # via aiohttp
beautifulsoup4==4.9.3     # via -r requirements.in
cachecontrol==0.12.6      # via firebase-admin
cachetools==4.1.1         # via google-auth
certifi==2020.11.8        # via httpx, requests
cffi==1.14.4              # via google-crc32c
chardet==3.0.4            # via aiohttp, httpx, requests
discord-py-slash-command==1.1.2  # via -r requirements.in
discord-py==1.7.1         # via -r requirements.in
firebase-admin==4.5.3     # via -r requirements.in
forex-python==1.5         # via -r requirements.in
future==0.18.2            # via python-twitter
google-api-core[grpc]==1.23.0  # via firebase-admin, google-api-python-client, google-cloud-core, google-cloud-firestore
google-api-python-client==1.12.8  # via firebase-admin, pydrive
google-auth-httplib2==0.0.4  # via google-api-python-client
google-auth==1.23.0       # via google-api-core, google-api-python-client, google-auth-httplib2, google-cloud-storage
google-cloud-core==1.4.3  # via google-cloud-firestore, google-cloud-storage
google-cloud-firestore==2.0.1  # via firebase-admin
google-cloud-storage==1.33.0  # via firebase-admin
google-crc32c==1.0.0      # via google-resumable-media
google-resumable-media==1.1.0  # via google-cloud-storage
googleapis-common-protos==1.52.0  # via google-api-core
googletrans==3.0.0        # via -r requirements.in
grpcio==1.33.2            # via google-api-core
h11==0.9.0                # via httpcore
h2==3.2.0                 # via httpcore
hpack==3.0.0              # via h2
hstspreload==2020.11.21   # via httpx
httpcore==0.9.1           # via httpx
httplib2==0.19.0          # via google-api-python-client, google-auth-httplib2, oauth2client
httpx==0.13.3             # via googletrans
humanfriendly==9.1        # via -r requirements.in
hyperframe==5.2.0         # via h2
idna==2.10                # via httpx, requests, yarl
instaloader==4.7.1        # via -r requirements.in
lxml==4.6.3               # via -r requirements.in
msgpack==1.0.0            # via cachecontrol
multidict==4.7.6          # via aiohttp, yarl
numpy==1.19.4             # via timezonefinder
oauth2client==4.1.3       # via pydrive
oauthlib==3.1.0           # via requests-oauthlib
proto-plus==1.11.0        # via google-cloud-firestore
protobuf==3.14.0          # via google-api-core, googleapis-common-protos, proto-plus
pyasn1-modules==0.2.8     # via google-auth, oauth2client
pyasn1==0.4.8             # via oauth2client, pyasn1-modules, rsa
pycountry==20.7.3         # via -r requirements.in
pycparser==2.20           # via cffi
pydrive==1.3.1            # via -r requirements.in
pyparsing==2.4.7          # via httplib2
pyreadline==2.1           # via humanfriendly
python-dateutil==2.8.1    # via -r requirements.in
python-twitter==3.5       # via -r requirements.in
pytz==2021.1              # via -r requirements.in, google-api-core, google-cloud-firestore
pyyaml==5.4               # via pydrive
requests-oauthlib==1.3.0  # via python-twitter
requests==2.25.0          # via cachecontrol, forex-python, google-api-core, google-cloud-storage, instaloader, python-twitter, requests-oauthlib
rfc3986==1.4.0            # via httpx
rsa==4.6                  # via google-auth, oauth2client
simplejson==3.17.2        # via forex-python
six==1.15.0               # via google-api-core, google-api-python-client, google-auth, google-auth-httplib2, google-resumable-media, grpcio, oauth2client, protobuf, python-dateutil
sniffio==1.2.0            # via httpcore, httpx
soupsieve==2.0.1          # via beautifulsoup4
timezonefinder==5.2.0     # via -r requirements.in
typing-extensions==3.7.4.3  # via aiohttp
uritemplate==3.0.1        # via google-api-python-client
urllib3==1.26.4           # via requests
yarl==1.5.1               # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
